+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
-## [Unreleased]
+## [0.6.2] 07-07-2020
+### Fixed
+* Fixed a bug in the script compiler that was not able to fetch the latest release ID since scripts are compiled BEFORE connecting to the server. The release ID is now passed into the compiler.
+
 ### Changed
 * The `install` command will now attempt to create the project root directory with a `mkdir -p` command.